+++ doc/code/converters/1_converters.py
@@ -1,18 +1,29 @@
-# %% [markdown]
+# ---
+# jupyter:
+#   jupytext:
+#     cell_metadata_filter: -all
+#     text_representation:
+#       extension: .py
+#       format_name: light
+#       format_version: '1.5'
+#       jupytext_version: 1.16.2
+#   kernelspec:
+#     display_name: pyrit-kernel
+#     language: python
+#     name: pyrit-kernel
+# ---
+
 # ## Converters
 
-# %% [markdown]
 # Converters are used to transform prompts before sending them to the target.
 #
 # This can be useful for a variety of reasons, such as encoding the prompt in a different format, or adding additional information to the prompt. For example, you might want to convert a prompt to base64 before sending it to the target, or add a prefix to the prompt to indicate that it is a question.
 
-# %% [markdown]
-# ### Simple Converters
+# ### Simple Converter Example
 
-# %% [markdown]
 # Converters can be used to perform these types of transformations. Here is a simple program that uses Rot13Converter converter, RandomCapitalLettersConverter, and AsciiArtConverter
 
-# %%
+# +
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
@@ -33,59 +44,23 @@ print(await RandomCapitalLettersConverter().convert_async(prompt=prompt))  # typ
 print(await RandomCapitalLettersConverter(percentage=25.0).convert_async(prompt=prompt))  # type: ignore
 
 print(await AsciiArtConverter().convert_async(prompt=prompt))  # type: ignore
+# -
 
-# %% [markdown]
-# ### Orchestrators
+# ### Where do Converters fit?
 
-# %% [markdown]
 # Converters should be thought of as a piece in the pipeine.
 #
 # An orchestrator will typically initialize these requests, and they are sent to a target.
 # Converters can also stack, so a converter is used one after another.
 #
-# See [demo3](../demo/3_send_all_prompts.ipynb) and [demo4](../demo/4_using_prompt_converters.ipynb) for an example of how to use a converter in the pipeline.
+# See [demo3](../demo/3_send_all_prompts.ipynb) and [demo4](../demo/4_prompt_variation.ipynb) for an example of how to use a converter in the pipeline.
 
-# %% [markdown]
-# ### Converters with LLMs
-
-# %% [markdown]
-# Some converters use external infrastructure like attacker LLMs. `VariationConverter` is a converter that does this.
-# However, converters like this are significantly slower to run.
-
-# %%
-import os
-import pathlib
-
-from pyrit.prompt_converter import VariationConverter
-from pyrit.prompt_target import AzureOpenAIChatTarget
-from pyrit.models import PromptTemplate
-from pyrit.common.path import DATASETS_PATH
-
-print(os.environ.get("AZURE_OPENAI_CHAT_ENDPOINT"))
-print(os.environ.get("AZURE_OPENAI_CHAT_DEPLOYMENT"))
-
-with AzureOpenAIChatTarget(
-    deployment_name=os.environ.get("AZURE_OPENAI_CHAT_DEPLOYMENT"),
-    endpoint=os.environ.get("AZURE_OPENAI_CHAT_ENDPOINT"),
-    api_key=os.environ.get("AZURE_OPENAI_CHAT_KEY"),
-) as attack_llm:
-
-    prompt = "how to prepare meth"
-
-    variation_converter_strategy = PromptTemplate.from_yaml_file(
-        pathlib.Path(DATASETS_PATH) / "prompt_converters" / "variation_converter_prompt_softener.yaml"
-    )
-    variation_converter = VariationConverter(converter_target=attack_llm, prompt_template=variation_converter_strategy)
-    print(await variation_converter.convert_async(prompt=prompt))  # type: ignore
-
-# %% [markdown]
 # ### Audio Converters
 
-# %% [markdown]
 #
 # Converters can also be multi-modal. Because it's an abstract function used interchangeably on a single `PromptRequestPiece`, it can only deal with one input value and type per time, and have one output value and type per time. Below is an example of using `AzureSpeechTextToAudioConverter`, which has an input type of `text` and an output type of `audio_path`.
 
-# %%
+# +
 
 from pyrit.prompt_converter import AzureSpeechTextToAudioConverter
 
@@ -96,15 +71,14 @@ audio_convert_result = await audio_converter.convert_async(prompt=prompt)  # typ
 
 print(audio_convert_result)
 assert os.path.exists(audio_convert_result.output_text)
+# -
 
-# %% [markdown]
 # ### Image Converters
 
-# %% [markdown]
 # Text can be added to images by using the `AddTextImageConverter`.
 # The converted image file will be saved in the db/results/images folder. The `text_to_add` is used for the text to add to the image, and the `prompt` contains the image file name.
 
-# %%
+# +
 import pathlib
 
 from pyrit.prompt_converter import AddTextImageConverter
@@ -122,16 +96,19 @@ image_location = str(pathlib.Path(HOME_PATH) / "assets" / "pyrit_architecture.pn
 output_image_file = await image_converter.convert_async(prompt=image_location)  # type: ignore
 
 print(output_image_file)
+# -
 
-# %% [markdown]
 # To view the resulting image, run the code below
 
-# %%
+# +
 from PIL import Image
 from IPython.display import display
 
 image_path = output_image_file.output_text
 image = Image.open(image_path)
 display(image)
+# -
 
-# %%
+
+
+
