+++ doc/demo/4_using_prompt_converters.py
@@ -3,7 +3,7 @@
 #
 # This demo shows how to use prompt converters in action.
 #
-# [Prompt Converters](../code/converters.ipynb) can be used to transform a prompt before they go to a target. They can be stacked, use LLMs, and are a powerful tool.
+# [Prompt Converters](../code/converters/1_converters.ipynb) can be used to transform a prompt before they go to a target. They can be stacked, use LLMs, and are a powerful tool.
 #
 # In all of these examples, NopTargets are used so these prompts are simply printed and added to memory. This can be useful if you are red teaming something and need to manually enter prompts. However, the target can be replaced with any other target. E.g., if you have API access you can add a target there.
 #
